backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
python-dotenv
google-generativeai
